+++ app/Http/Controllers/Api/OrgController.php
@@ -12,6 +12,8 @@ class OrgController extends Controller
 {
     public function index()
     {
+        $this->authorize('viewAny', Org::class);
+
         $orgs = Org::paginate(15);
 
         return OrgResource::collection($orgs)->response()->setStatusCode(200);
@@ -19,6 +21,8 @@ class OrgController extends Controller
 
     public function store(OrgStoreRequest $request)
     {
+        $this->authorize('create', Org::class);
+
         $org = Org::create($request->validated());
 
         return new OrgResource($org);
@@ -27,9 +31,11 @@ class OrgController extends Controller
     public function show(string $id)
     {
         $org = Org::find($id);
-        if (!$org) {
+        if (! $org) {
             return response()->json(['message' => 'Not Found'], 404);
         }
+
+        $this->authorize('view', $org);
 
         return new OrgResource($org);
     }
@@ -37,9 +43,11 @@ class OrgController extends Controller
     public function update(OrgUpdateRequest $request, string $id)
     {
         $org = Org::find($id);
-        if (!$org) {
+        if (! $org) {
             return response()->json(['message' => 'Not Found'], 404);
         }
+
+        $this->authorize('update', $org);
 
         $org->update($request->validated());
 
@@ -49,9 +57,11 @@ class OrgController extends Controller
     public function destroy(string $id)
     {
         $org = Org::find($id);
-        if (!$org) {
+        if (! $org) {
             return response()->json(['message' => 'Not Found'], 404);
         }
+
+        $this->authorize('delete', $org);
 
         $org->delete();
